Split out users + stardardize file names
