+++ src/server/implementation/types.ts
@@ -3,6 +3,7 @@ import {
   FunctionReference,
   GenericActionCtx,
   GenericMutationCtx,
+  GenericQueryCtx,
   TableNamesInDataModel,
   defineSchema,
   defineTable,
@@ -121,6 +122,7 @@ const defaultSchema = defineSchema(authTables);
 export type AuthDataModel = DataModelFromSchemaDefinition<typeof defaultSchema>;
 export type ActionCtx = GenericActionCtx<AuthDataModel>;
 export type MutationCtx = GenericMutationCtx<AuthDataModel>;
+export type QueryCtx = GenericQueryCtx<AuthDataModel>;
 export type Doc<T extends TableNamesInDataModel<AuthDataModel>> = GenericDoc<
   AuthDataModel,
   T
